requirements.txt
tensorflow
numpy
matplotlib
opencv-python
streamlit
scikit-learn
google-generativeai
seaborn
plotly
pandas
pillow
